LayoutTests/http/tests/referrer-policy-script/no-referrer/same-origin.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests the behavior of no-referrer referrer policy when same origin.");
jsTestIsAsync = true;

function checkReferrer(value) {
    referrer = value;
    // Should be the empty string
    shouldBeEqualToString("referrer", "");
    finishJSTest();
}

if (window.testRunner) {
    testRunner.setStatisticsShouldDowngradeReferrer(false, function () { downgradeReferrerCallback('no-referrer', '127.0.0.1')});
}
</script>
</body>
</html>
